CSS, JS and IMG files
Created the folder "css" in "public" and inside it the "styles.css" file to use to edit the styles of the site.

Created the folder "img" in "public" and inside it was put the imgs there will be used in the site.

Created the folder "js" in "public" and inside it the "scripts.js" file to use to write the JavaScript's scripts of the site.